refaktorerer journalføringsfiler inn i mapper. Forberedelse til journalføring 2.0
