+++ src/frontend/Komponenter/Admin/AdminApp.tsx
@@ -1,7 +1,7 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
 import React from 'react';
-import { JournalføringAdmin } from '../Journalføring/JournalføringAdmin';
-import { JournalføringAdminVelger } from '../Journalføring/JournalføringAdminVelger';
+import { JournalføringAdmin } from '../Journalføring/Admin/JournalføringAdmin';
+import { JournalføringAdminVelger } from '../Journalføring/Admin/JournalføringAdminVelger';
 import GamleBehandlinger from '../Behandling/GamleBehandlinger/GamleBehandlinger';
 
 export const AdminApp: React.FC = () => {
