+++ src/frontend/Komponenter/Journalføring/Admin/JournalføringAdmin.tsx
@@ -1,26 +1,26 @@
 import React, { useEffect, useState } from 'react';
 import { Navigate, useNavigate, useParams } from 'react-router-dom';
-import { RessursFeilet, RessursStatus, RessursSuksess } from '../../App/typer/ressurs';
+import { RessursFeilet, RessursStatus, RessursSuksess } from '../../../App/typer/ressurs';
 import styled from 'styled-components';
-import Brukerinfo from './Brukerinfo';
-import { stønadstypeTilTekst } from '../../App/typer/behandlingstema';
-import DataViewer from '../../Felles/DataViewer/DataViewer';
-import { useHentJournalpost } from '../../App/hooks/useHentJournalpost';
-import { useHentFagsak } from '../../App/hooks/useHentFagsak';
-import { useApp } from '../../App/context/AppContext';
-import { Behandlingstype, behandlingstypeTilTekst } from '../../App/typer/behandlingstype';
-import { utledRiktigBehandlingstype } from './journalførBehandlingUtil';
+import Brukerinfo from '../Felles/Brukerinfo';
+import { stønadstypeTilTekst } from '../../../App/typer/behandlingstema';
+import DataViewer from '../../../Felles/DataViewer/DataViewer';
+import { useHentJournalpost } from '../../../App/hooks/useHentJournalpost';
+import { useHentFagsak } from '../../../App/hooks/useHentFagsak';
+import { useApp } from '../../../App/context/AppContext';
+import { Behandlingstype, behandlingstypeTilTekst } from '../../../App/typer/behandlingstype';
 import {
     hentFraLocalStorage,
     lagreTilLocalStorage,
     oppgaveRequestKey,
-} from '../Oppgavebenk/oppgavefilterStorage';
-import { IJojurnalpostResponse, journalstatusTilTekst } from '../../App/typer/journalføring';
-import { formaterIsoDatoTid } from '../../App/utils/formatter';
-import { UtledEllerVelgFagsak } from './UtledEllerVelgFagsak';
-import { AlertError, AlertWarning } from '../../Felles/Visningskomponenter/Alerts';
+} from '../../Oppgavebenk/oppgavefilterStorage';
+import { IJojurnalpostResponse, journalstatusTilTekst } from '../../../App/typer/journalføring';
+import { formaterIsoDatoTid } from '../../../App/utils/formatter';
+import { UtledEllerVelgFagsak } from '../Felles/UtledEllerVelgFagsak';
+import { AlertError, AlertWarning } from '../../../Felles/Visningskomponenter/Alerts';
 import { Button, Heading } from '@navikt/ds-react';
-import { BodyLongSmall, BodyShortSmall } from '../../Felles/Visningskomponenter/Tekster';
+import { BodyLongSmall, BodyShortSmall } from '../../../Felles/Visningskomponenter/Tekster';
+import { utledRiktigBehandlingstype } from '../Felles/utils';
 
 const Blokk = styled.div`
     margin-bottom: 1rem;
