+++ src/frontend/Komponenter/Journalføring/Felles/VelgFagsakForIkkeSøknad.tsx
@@ -1,9 +1,9 @@
-import { IJojurnalpostResponse } from '../../App/typer/journalføring';
+import { IJojurnalpostResponse } from '../../../App/typer/journalføring';
 import {
     behandlingstemaTilStønadstype,
     Stønadstype,
     stønadstypeTilTekst,
-} from '../../App/typer/behandlingstema';
+} from '../../../App/typer/behandlingstema';
 import React, { useEffect, useState } from 'react';
 import { Select } from '@navikt/ds-react';
 import styled from 'styled-components';
