+++ src/frontend/Komponenter/Journalføring/Klage/BehandlingKlageInnold.tsx
@@ -1,13 +1,13 @@
 import React, { Dispatch, SetStateAction, useEffect, useState } from 'react';
 import styled from 'styled-components';
-import DataViewer from '../../Felles/DataViewer/DataViewer';
-import { formaterIsoDatoTid, formaterNullableIsoDato } from '../../App/utils/formatter';
-import { Ressurs, RessursStatus } from '../../App/typer/ressurs';
-import { KlageBehandling, Klagebehandlinger } from '../../App/typer/klage';
-import { BehandlingKlageRequest } from '../../App/hooks/useJournalføringKlageState';
+import DataViewer from '../../../Felles/DataViewer/DataViewer';
+import { formaterIsoDatoTid, formaterNullableIsoDato } from '../../../App/utils/formatter';
+import { Ressurs, RessursStatus } from '../../../App/typer/ressurs';
+import { KlageBehandling, Klagebehandlinger } from '../../../App/typer/klage';
+import { BehandlingKlageRequest } from '../../../App/hooks/useJournalføringKlageState';
 import { Alert, Button, Checkbox, Heading } from '@navikt/ds-react';
 import { PlusCircleIcon } from '@navikt/aksel-icons';
-import { harÅpenKlage } from '../../App/utils/klage';
+import { harÅpenKlage } from '../../../App/utils/klage';
 
 interface Props {
     settBehandling: (behandling?: BehandlingKlageRequest) => void;
